Refactor `Bridge` to `Relay`
This commit refactors the implementation of `Bridge` to `Relay`. While
the concept is largely the same, the API to use it differs
significantly.

Previously with `Bridge`, it was supposed to be a single interface into
external services, however upon careful consideration it is apparent
that these `Bridges` are nothing but `Producer`s who understand external
services.

Thus, this commit removes the opinionated `Bridge` concept into a
generic `Relay` concept, which means now clients can easily write their
own version of a `Relay`, without having to first merge it upstream.

The README is also updated accordingly.
