+++ lib/rocketman.rb
@@ -4,4 +4,3 @@ require 'rocketman/registry.rb'
 require 'rocketman/event.rb'
 require 'rocketman/producer.rb'
 require 'rocketman/consumer.rb'
-require 'rocketman/bridge.rb'
